moves desktop as user logic to own pkg for darwin
