+++ pkg/runas/runas_darwin.go
@@ -1,7 +1,7 @@
 //go:build darwin
 // +build darwin
 
-package runner
+package runas
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 )
 
 // For notifications to work, we must run in the user context with launchctl asuser.
-func (r *DesktopUsersProcessesRunner) runAsUser(_ context.Context, uid string, cmd *exec.Cmd) error {
+func RunAsUser(_ context.Context, uid string, cmd *exec.Cmd) error {
 	// Ensure that we handle a non-root current user appropriately
 	currentUser, err := user.Current()
 	if err != nil {
